Changelog:
Added cloth simulation to main project. While during the descending phase of the cloth the simulation looks real and appropriate, once reached the "rest" position the particle-spring lattice just do not behave as I want to. While messing with input data, I was able to find the source of the problem but not manage to find a proper solution to it (from line 356 to 366).